+++ .idea/jarRepositories.xml
@@ -69,22 +69,27 @@
     <remote-repository>
       <option name="id" value="WPILocal" />
       <option name="name" value="WPILocal" />
-      <option name="url" value="file:$USER_HOME$/wpilib/2022/maven" />
+      <option name="url" value="file:$PROJECT_DIR$/../wpilib/2022/maven" />
     </remote-repository>
     <remote-repository>
       <option name="id" value="WPILocal" />
       <option name="name" value="WPILocal" />
-      <option name="url" value="file:$USER_HOME$/wpilib/2023/maven" />
+      <option name="url" value="file:$PROJECT_DIR$/../wpilib/2023/maven" />
     </remote-repository>
     <remote-repository>
       <option name="id" value="WPILocal" />
       <option name="name" value="WPILocal" />
-      <option name="url" value="file:$USER_HOME$/wpilib/2023/maven/" />
+      <option name="url" value="file:$PROJECT_DIR$/../wpilib/2023/maven/" />
     </remote-repository>
     <remote-repository>
       <option name="id" value="WPILocal" />
       <option name="name" value="WPILocal" />
       <option name="url" value="file:/$PROJECT_DIR$/../../../Public/wpilib/2023/maven" />
+    </remote-repository>
+    <remote-repository>
+      <option name="id" value="WPILocal" />
+      <option name="name" value="WPILocal" />
+      <option name="url" value="file:/$PROJECT_DIR$/../../Public/wpilib/2023/maven/" />
     </remote-repository>
   </component>
 </project>
